lava-templates/fvp-*: Decrease test timeout to 12min
This timeout kicks in when the test fails expectations (i.e. it keeps
waiting for a particular output, but as it doesn't arrive, it keeps
waiting until timeout). Recently we had a case when a lot of test jobs
started to fail like that at the same time (apparently because of influx
of new patches), which caused a bottleneck in LAVA queue, which in turn
lead to domino effect on Jenkins side, which jobs bottlenecking and
timing out.

Based on job log, the maximum runtime we have is 10-11 minutes. So,
cut down the timeout to 12 min (25% decrease), which should alleviate
bottlenecking issue at least somewhat, until a more flexible solution
is found.

Related to https://linaro.atlassian.net/browse/TFC-368

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I6738ae266a725b7399ab8564141ce77c0183c880
(cherry picked from commit 9bf5108f4e7f150439fd07b8756506f0e46b83df)