scp: replace deprecated makefile command
Previously `make clean` as used to clean the SCP-Firmware workspace.
With the deprecation of `Makefile` this command will not longer work
for that reason it was replaced with the corresponding one that
uses the `Makefile.cmake` wrapper file.

Signed-off-by: Leandro Belli <leandro.belli@arm.com>
Change-Id: I0fcf091aef345ebb5b24fd3aef06c4e15717b61c
(cherry picked from commit 99e20b29635fe19bcc8b11199f5f44db392c466a)