fix: specify FVP start ports explicitly
This change ensures that every supported FVP is explicit about the ports
that it will try to bind to. This is a mitigation against the
sometimes-random order in which terminals are enumerated by the various
FVPs, which causes our LAVA tests to attempt to read from the wrong
terminal.

Signed-off-by: Chris Kay <chris.kay@arm.com>
Change-Id: Ida12fcdae988fe119226c05032c90387e574727b
(cherry picked from commit 30120b38df617c05307ee42f46bc6f5db5d7f568)