fvp_utils.sh: If "failures" pattern used in expect-lava, make them fatal
"Failures" pattern in:

expect_string+=("i;<prompts>[;<successes>;<failures>;<commands>]")

usually not specified (lack of success is considered a failure), but if
it's specified, consider it fatal, i.e. fail-fast the entire test job
(by raising an exception). Otherwise, only a particular testcase is
recorded as failed, and the rest of the job is continued.

This comes from the https://linaro.atlassian.net/browse/TFC-442 report,
where a bootloop produced 200+MB logs, because the job just continued
to run and capture DUT output until it timed out.

Arguably, this fail-fast behavior for "failures" is what we need all of
the time, so just make it default. (If later it is found example of
contrary, we would need to introduce flags to the expect_string syntax
for finer control).

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I57576505a38357bef8521bbc3c8b4f119d508a30
(cherry picked from commit 5b5eadaf056985764732de29e48314f5494b67ed)