feat(tfut): support unit testing build config
By adding support for the unit testing build config in the test
group, the CI will be able to build the unit tests based on that
config.

We are not targeting specific groups yet. There isn't anything
that checks whether it is of a specific group of tests and transforms
the config. This will be addressed in a further patch.

Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>
Signed-off-by: Edward Potapov <edward.potapov@arm.com>
Change-Id: I79f4a356c183d29d3758759fe1a54596c57fe536