fix: check and initialise spm submodules
As part of the tf-a-job-configs we may clone the spm module to sync
it to the other repos based on the topic however we don't clone
the submodules at this stage. Therefore even if the spm directory
exist we still want to clone the submodules. This patch makes the
checks if the submodules have been initialsied and if not does it.

Signed-off-by: Daniel Boulby <daniel.boulby@arm.com>
Change-Id: If07fb5c721b699bf80bd7fe5f54fad5ad16ed33d
(cherry picked from commit 25385ab4ed41e89ae8c61e5db2f2607074f6a7f7)