feat(tfut): add support for hooks and run configs
With this patch, the pre_tfut_build hook can be implemented in the
TFUT run config files in order to perform custom actions, such as
setting build targets. This way, build targets can be specified in
the TFUT run config files, as it is currently done for TF-A, allowing
for more flexibility.

Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>
Signed-off-by: Edward Potapov <edward.potapov@arm.com>
Change-Id: If8a4d0c63a885eee9ffdd8f2c5a68c6a3cff9cd4