gpt: enhance gen_gpt_bin function to accept max. FIP image size
Updated GPT image creation function to now accept the size
of the FIP image as an input parameter.
Furthermore, streamlined and enhanced the GPT creation function
to offer more detailed information about the process of GPT image
creation.

This update proves beneficial for creating a partition for the FIP,
allowing us the flexibility to choose the desired size rather
than adhering to a fixed size.

Signed-off-by: Manish V Badarkhe <Manish.Badarkhe@arm.com>
Change-Id: I1e27bf48c1bae0dfdc8d012d85c613852fd30cd4