feat(tfut): handle test interruption properly
When tests are run locally and the execution is interrupted by the
user, if tests keep running (e.g.: infinite loops, models not
terminating properly), the script keeps waiting for test
termination, as it receives the interruption signal (e.g. Ctrl+C)
but waits for the test to terminate, which may never happen.

This patch addresses that issue by properly addressing test
termination upon interruption or script failure.

Signed-off-by: Juan Pablo Conde <juanpablo.conde@arm.com>
Signed-off-by: Edward Potapov <edward.potapov@arm.com>
Change-Id: I1739f3a95736bb65d93e7b63d765cf345cb75250