feat: drop spm-uart2.exp
The first cactus SP doesn't own an UART device.
As such, its logs will be mixed with tftf and the other
partitions.

The spm-uart2.exp was only checking that first cactus has booted.
This was when it was assigned with its own uart device.

For validating that the SP1 was booted, we have the use of
FFA_PARTITION_INFO_GET to check it's present in the running
system. This is done by TFTF at most SPM-related tests that need
to interact with the SP.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: Ifb8433c7438498f6346e89d89a85bcbfb998da31