tf-l3-boot-tests-css: Remove TC1/TC2 configs
These configs fail to boot with LTS 2.10 branch ("Kernel panic - not syncing:
VFS: Unable to mount root fs on unknown-block(0,0)"), at the very least
because proper kernel params are not passed. The builds for these configs
depend on external prebuilt files, which are clearly not versioned for the
LTS releases, and are thus "owned" by the mainline and potentially changed
on its needs at any time. The current status in the mainline is that support
for TC1 is removed, as that platform is deprecated, while TC2 support has
undergone extensive changes via multiple commits in tf-a-ci-scripts. My
understanding is also that TC1/TC2 are "virtual" IP platforms, available
via FVP, to let partners to develop integration of it into specific
silicon products, while the focus of the TF-A LTS appears to be the actual
supported silicon products, to allow to deliver security and other
important fixes to the market products.

Thus, between backporting TC2 support from mainline (and not being clear
what to do with TC1 support), vs just removing those 2 broken configs
to unbreaks LTS2.10 CI, the latter appears to be more reasonable choice
as of now. If at later time it is decided that TC1/TC2 need to be
supported, this change can be reverted and detailed engineering effort
applied towards upgrading it.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: Ia213695824ef8923edd2ca4e924e44b10ad4e4a1