ci(cleanup): remove aemv8a run from group if aemva is present
This is first patch in series of reducing CI run time without
compromising on code-coverage by removing duplicate CIs in same group
which run for both aemva and aemv8a.

FVP_AEMVA and FVP_AEMV8A are same models and it is just a name change
to cover Armv9 models as well. The code coverage with and without the
patch are same.

Signed-off-by: Manish Pandey <manish.pandey2@arm.com>
Change-Id: If515b9d8644cb77f38924941a2a2323932de6647