ci: remove unwanted configs from extensive tests
As per CI clean-up process we are reducing the number of duplicate test
configuration which does not bring any value to the project. One of the
example is to remove cortexa35x4 and cortexa57x4a53x4-tspd from each
group.
Also, for extensive PSCI tests singlecore does not bring any real value
as these tests are skipped if number of core is less than two.

While at this, use aemva instead of aemv8a for default test.

Change-Id: Ib78d20037ccdeb4e6f76ca1f25ea5690f0e7c98b
Signed-off-by: Manish Pandey <manish.pandey2@arm.com>