tc2: add mte2 for tc2 builds
The kernel image used to boot tc2 in CI expects MTE2 to be enabled.
In TF-A SCR-EL3.ATA bit(26) was set without any dependency,
However with patch:
https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/26891

SCR-EL3.ATA depends on FEAT_MTE2, since we are using a kernel image
which expects the SCR-EL3.ATA to be set, we need to enable FEAT_MTE2
for kernel boot tests with tc2.

Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>
Change-Id: I1ff4b68f90b5e03c1cbfa75af56dc5260de90ac2