fix: temporarily drop two spm-l3-boot-tests
Dropping tests that have run config that depends on
SPM artifacts:
- fvp-spm.linux
- fvp-spm.rstbl31

The first is using an outdate setup. Team will look to provide
a more recent and meaningful test.

The latter validates the reset to BL31 option, and is validated in
the hafnium tests. This setup can be updated, similarly to the above.

Signed-off-by: J-Alves <joao.alves@arm.com>
Change-Id: I701e5974a9bd0d98d70302c177b1ea187faa4e20