Revert "lava-templates: Comment out "license_variable" param"
This reverts commit 1293a1412661fb1d48a0fe444efaa8860db6f9e6.

Reason for revert: TF-A CI uses non-UBL-enabled FVP models which
are hard to upgrade. All relevant parties were notified and aware
of the issue, but work on that may conflict with the preparation
of the upcoming release. We are starting with the contingency plan
of still extending FlexNet licenses expiration. The matter to be
revisited in June 2024

More info in https://linaro.atlassian.net/browse/TFC-608

Change-Id: I3d092e7558dc84dc2683ac30258a32b258896b0c