ci(errata_abi): remove Cortex-A710 test
Currently the errata ABI is tested in different configurations
out of which we have Cortex-A710 FVP check this configuration
though it uses an older 11.17 model.

This configuration build is on image size boundary and adding
anything to BL31 will cause this build to fail, but
unfortunately we cant set SRAM size to 384KB on this configuration
since this is on an older 11.17 model that doesn't support bigger
SRAM. Trying to move this test to newer model doesn't work for some
strange reason on CI but works locally.

Links to test run fails with the newer models below:
11.22.14 - https://ci.trustedfirmware.org/job/tf-a-ci-gateway/69894/
11.25.20 - https://ci.trustedfirmware.org/job/tf-a-ci-gateway/70014/

So, until we find the reasoning behind this, remove this test
configuration, as it is not a critical configuration and currently
blocks any addition to BL31. For example, the current updates
to BL31 with SMCCC-1.5 are blocked."

Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>
Change-Id: I6f91e29a8f5bb4e998f7b562e5b5464b4397fd82