fix: disable debug builds for mbedtls dualroot
Migrating from 2.28.5 to 2.28.8 mbedtls involves a bit of a size increase causing certain FVP build configurations to increase in size.
So disable debug builds to workaround this problem.

This may be a FVP specific problem as we build TF-A for numerous CPUs all together and hence should not be a problem for individual platforms as the number of CPUs are limited in that case.

Change-Id: I5e63329d9df6cd86e3ccba97c4a47d813e44d9fe
Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>