fix(coverity): add additional trap information
Errors in the coverity scripts are handled silently, to ensure that we
get results for all configurations. This makes it hard, in certain
instances, to know where a build failure has occured and why. Add
additional information to the trap handler to make it easier to resolve
hard to debug errors.

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: I762f04c2058be18a60f94ec77a921b6a5770a2ac