ci(cleanup): remove few tests from tftf-l2-fvp-dynamiq
This patch removes a few tests from the tftf-l2-fvp-dynamiq
test group that are duplicated or covered by similar platforms
or test configs.

Ensured to make sure that the test group covers
tests for atleast one CPU from the fvp-makefile under the build
flags HW_ASSISTED_COHERENCY=0 or CTX_INCLUDE_AARCH32_REGS=0 or
when both the above flags are enabled.

List of tests that were retained are:

1. fvp-dynamiq-aarch64-only,fvp-default:fvp-dtb.dyn-tftf-fip.tftf-neoversev1x4-debug
2. fvp-dynamiq-aarch64-only,fvp-default:fvp-dtb.dyn-tftf-fip.tftf-cortexa76aex4-debug
3. fvp-dynamiq-aarch64-only,fvp-default:fvp-dtb.dyn-tftf-fip.tftf-cortexa710x4-debug
4. fvp-dynamiq-aarch64-only,fvp-default:fvp-dtb.dyn-tftf-fip.tftf-neoverse_n2x4-debug
5. fvp-dynamiq-aarch64-only,fvp-default:fvp-dtb.dyn-tftf-fip.tftf-cortexx2-debug

Signed-off-by: Sona Mathew <sonarebecca.mathew@arm.com>
Change-Id: I10dc1a8f5d23f3d0088880408d95fa6ef74de969