static-checks-coding-style.sh: Make it actually check the last patch
This script is documented as "Check coding style on the last patch", but
doesn't work like that and actually runs on already existinf commits on
the branch too. If some of those commits were merged with codestyle
exceptions, this leads to false positive codestyle errors, completely
unrelated to the current patch.

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: Ibf5763fdeecf2bacc64f0bf215dfbbcf31b3e480