static-checks-coding-style.sh: Check entire patch chain
Due to the way LTS vote propagation works, we actually need to test the
entire patch chain. The process is: in case of patch chain, Allow-CI is
set on the final patch in chain, other patches don't really get Allow-CI
votes. So, CI on the final patch should test the entire series, set
Verified vote on it, then this vote is propagated to rest of patches in
series.

Consequently, the changes are:

1. Proper branch should be used to find patch series limit (instead of
hardcoding master).
2. Diagnostics for codestyle checks (and actually all the other CI tests)
should make it clear that it checks entire series (and not just "last
patch").

Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>
Change-Id: I803a745f7219ce4d0102027f026346a93261ee49