tc: Add support for TC3 platform
This includes scp-boot-tests, tf-a plat test, tf-a boot test, and spm-l3
boot tests.

As time being when submitting this patch, the measured boot and DPE are
not supported on TC3 yet, therefore, the measured boot and DPE test
scripts are not included.

We will introduce these tests on TC3 later when the functionalities are
ready.

Change-Id: I087cbb826cfc39d49614e0cc250f0867d06b204f
Signed-off-by: Quoc Khanh Le <QuocKhanh.Le@arm.com>
Signed-off-by: Leo Yan <leo.yan@arm.com>