fvp: move a710 to 8 core test
With 11.26.11 FVP models a710 is a 8 core FVP, so move tests
to use 8 cores rather than 4 cores.

If we setup as 4 cores from FVP kernel still recognises as 8 core system.
[...]
 rcu: 	RCU restricting CPUs from NR_CPUS=256 to nr_cpu_ids=8.
[...]

This means kernel can schedule init.sh on any of the core ids.
But since we fail to bring up core4 to 7 and we schedule on any of
these cores will eventually lead to a hang with init.sh
not mounting the rootfs, if init.sh is run on core4 to 7.

Also in ci scripts we wait for shell prompt "#" for test to pass
which may not happen when init.sh fails.

Change-Id: Iadd87cbbe8bbb2d456496b9ce65e785f686b66af
Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>
(cherry picked from commit 268a2befdb6d55a6e6b8347357b2b12cac580947)