tc: Remove rotpk test on TC2
The rotpk test is broken if we move to use the latest RSE firmware, as
the PAS function ID RSE_CRYPTO_EXPORT_PUBLIC_KEY_SID mismatches between
TF-A and RSE.

The rotpk test can make success on TC2 which is because CI test sticks
to an old RSE firmware so it has no mismatching issue.

We will enable the test on TC3 soon for the corrected
RSE_CRYPTO_EXPORT_PUBLIC_KEY_SID function ID, as a side effect, this
will lead to test failure on TC2 with the updated TF-A and the old RSE
binary.  On the other hand, currently we have no plan to update RSE
binary, so remove the rotpk test on TC2.

Change-Id: I427907a0df96ec3c520b11a37e95b86b4f08b4c3
Signed-off-by: Leo Yan <leo.yan@arm.com>