chore(rdfremont): rename RD-Fremont to RD-V3
Arm has decided to rename RD-Fremont to RD-V3 to align with its
existing product lineup, such as RD-V1, RD-V2, etc. This patch [1]
replaces all occurences of "Fremont" with "V3" in file names and
contents. This change updates the testing config name and build
config of the CI to match the above patch.

[1] https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/29845

Change-Id: I26b47f98d15542035ade5c1cddcbf28e2f309dee
Signed-off-by: Jerry Wang <Jerry.Wang4@arm.com>