tc: Fix up SCP binary path
On TC4, the SCP binary is unified to single image scp-css.bin, it does
not distinguish SCP BL1 and BL2 images. As a result, it is pointless to
test the SCP firmware with specifying SCP_BL2 option in fiptool.

This commit simply takes SCP binary as SCP ROM binary (scp-bl1.bin) to
fit into the CI framework. It skips to pass SCP_BL2 parameter for
building fip image on TC4.

Change-Id: Ia46bd8f855a16edc788106f7e0e5721582f0ca34
Signed-off-by: Leo Yan <leo.yan@arm.com>