fix(ci): fix dependency group name
The dependency group for documentation has been renamed to docs [1],
rename it to align with this change. This change in the CI scripts was
missed. In CI we are using Poetry 1.3.2 which elides over spelling
errors in the group name. However, in newer versions, incorrect group
names are treated as fatal errors by Poetry.

[1] https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/30344

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: I16cdb1538433cf4fffba0ee76c5e72327e05e401