refactor(cpu-extensions): drop explicit build and run configs
fvp-aemv8a.all already runs with MPAM enabled. The default FVP build has
MPAM=2 (inherited from the defaults) as well as MTE2=2. Both dedicated
tests those run with the same cpu-extensions tftf build. As such the
MPAM specific test case is entirely covered and redundant, while the
MTE2 test can be absorbed into the all config by adding MTE2 to the run
command and building with the tsp which is added to the build config.

Also the bmcov run config is unused so drop it.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: Iaa54a5e79bd3a56e4690580548a3f842e7682725