chore(u-boot): bump to version 2024.10
Upgrade U-Boot from 2019.07 to 2024.10 to support firmware handoff boot
tests, as many underlying assumptions have changed. Notably, newer
updates rely on semihosting on FVP instead of the previous approach.

The defconfig used to build the current U-Boot binary in CI was removed
some time ago [1]. Previously, U-Boot scanned DRAM for boot binaries.
This approach was replaced with semihosting, allowing the model to
locate boot images via the host filesystem.

To maintain compatibility and minimize changes in CI scripts, a boot
script [2] is provided to emulate this assumption. All our DTs assume
the root partition is /dev/vda, but the RootFS used in CI uses
/dev/vda2. Update it to avoid having to awkwardly fixup the location in
the DT in CI.

[1] https://lists.denx.de/pipermail/u-boot/2019-August/382053.html
[2] https://downloads.trustedfirmware.org/tf-a/linux_boot/fvp/boot.scr

Signed-off-by: Harrison Mutai <harrison.mutai@arm.com>
Change-Id: I3b05972c1a665bdd30206a7cab22d0ae4a32a8d6