fix: disable FEAT_FPMR for Undef injection test
Undef injection requires FEAT_FGT to be set to 0 for testing on FVP with
all features are set FEAT_STATE_CHECK (=2). After introduction of
FEAT_FPMR (which defaults to 2) it fails to build because FPMR has
dependency on FGT.

To solve the issue, disable FPMR on undef injection build configuration.

Change-Id: I4ae907c2558b8350a6e569ce303c6a8a6fc36f4c
Signed-off-by: Manish Pandey <manish.pandey2@arm.com>