test: remove optee spmc based configs
The test config [1] exercises OPTEE as S-EL1 SPMC on top of TF-A SPMD
using tftf to run tf-a-tests suites. It uses an old prebuilt binary of
OPTEE SPMC. This test config served a limited purpose of exercising the
FF-A virtual and physical interfaces at the time when it was introduced
in 2021.

Fast forward, the OPTEE CI has evolved to offer a much exhaustive test
coverage of FF-A interface between SPMC and SPMD. The above mentioned
test config does not offer any explicit benefit. Moreover, a recent
prebuilt of OPTEE SPMC cannot be used since there is a significant
deviation in functionality offered by OPTEE versus whats expected by
the tf-a-tests suite.

It is not worth the effort to tweak tf-a-tests driver to make OPTEE
SPMC run successfully.

[1] spm-l2-boot-tests/fvp-default,fvp-optee-as-spm,fvp-default:fvp-spm.optee

Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>
Change-Id: I1dbab6a216bc62c764712dc03cf4a543963a52cd