test: add config that tests DYNAMIC_WORKAROUND_CVE_2018_3639
Cortex-A76 is affected by CVE_2018_3639 but the issue has been
fixed on newer revisions through inclusion of FEAT_SSBS.
We check if FEAT_SSBS is implemented and apply the mitigation
only if it's not implemented. This feature check has been true
since r3p0 and we don't test the case where mitigation
is required in CI.

This patch allows us to test the alternate case for
improving coverage.

Signed-off-by: Arvind Ram Prakash <arvind.ramprakash@arm.com>
Change-Id: I53a5c81dc1e1ebd6eaca9f3bce5b81d9c6569b99