fix: boot a linux kernel as preloaded bl33 image
Fix the CI boot flow that incorrectly assumes a successful Linux
kernel boot when booting a preloaded kernel image.

To boot the kernel using initrd as a ram file system the kernel
requires the "initrd_start" and "initrd_end" properties in the
chosen node of the device tree. The "initrd_end" property can be
derived from the actual initrd file (using its size). The
"INITRD_PATH" build time variable will initiate the insertion of the
initrd properties in to the device tree. Additionally this variable
takes the location of an initrd file to determine its size.

Change-Id: Idd9b5f9e5f1142a8ed1b8901d6a7c0d095d4850a
Signed-off-by: Salman Nabi <salman.nabi@arm.com>