ci: extend the all feature configs with more features
Enable features that we know of but weren't enabled. Add support for
FEAT_MOPS as we have code to use it in tf-a. Factor out features that
were enabled only for rme but are generic.

Propagate all of this to the bmcov job as it's got out of date.

We can remove the explicit enables for ENABLE_FEAT_MTE2 as it's default
enabled on fvp.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: Ieda169bff6e50fae237f7e2e43d51fa9c1ab9cd6