fix(ci): downgrade ARM_ARCH_MAJOR and ARM_ARCH_MINOR on generic build configs
This directly propagates to the -march directive which enables the use
of architectural features. The fvp-spm-sve+sme build config may or may
not be run in a model with these features enabled. So don't specify
them so we don't get faults in random places.

But keep it at a reasonable minimum, so that we get BIT land pads.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I1668ba753a14149431bd8ed63e153e112056475e