fix(ci):add covetrace.so plugin on local ci run
Fix for run_config fragments to prevent covetrace plugin
errors. When running on local ci, the covtrace plugin will be
downloaded from tfa/downloads and set as bmcov plugin path to
mitigate the past errors of the plugin not being present.
The covetrace plugin is downloaded as the warehouse does not
host the plugin.

Signed-off-by: Thaddeus Gonzalez-Serna <Thaddeus.Gonzalez-Serna@arm.com>
Change-Id: Ie776d24079d787cd2331f1ac4acb6181ed38ba24