fvp: use -Og instead of -O0
The fvp is an unrealistic platform - by default it includes as many
things (features, cpus, etc) as possible. Building with -O0 makes the
BL31 binary immense. The fvp is also not unconstrained - the maximum
amount of SRAM it can have is 512K. So build with -Og instead. It is the
next closest to an unoptimised build without excessively increasing code
size. For detecting any actual -O0 problems, there is still the juno
build.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: Id1167ad4c76b4be10e73b41987dfe924f584c0dd