feat: add static checkers for cpu workaround order
Workarounds added to CPU files are expected to be in order
based on Erratum ID and also if the CVE's are added they should
be ordered by CVE Year and CVE ID.

There is no strict enforcement currently in place to check this
there is a runtime check for that in TF-A, but booting up
every FVP model for this would be too cumbersome.

Ref:
https://git.trustedfirmware.org/plugins/gitiles/TF-A/trusted-firmware-a/+/refs/heads/integration/lib/cpus/errata_report.c#102

So add a static checker to check in-order placements
of erratum and CVE's in CPU files.

Change-Id: Ia86de7a982e974d64564d09c950b4b680b4e0f7f
Signed-off-by: Govindraj Raja <govindraj.raja@arm.com>