fix(tfa-next): run FVP with Arm v8.4 extension emulation
RF-A requires at least Arm v8.4 extension for accessing EL2
GICv3-related registers. Right now, we do not specify any specific Arm
extension in the FVP run configuration file so the FVP will emulate
Arm v8.0 as a baseline. Consequently, RF-A crashes when trying to
write into the ICC_SRE_EL2 register.

We uncovered the same issue in the past and fixed it in the RF-A
makefile itself as part of RF-A commit 406eb9ec8edb72d "feat(rust): do
setup for EL2 GICv3-related registers for FVP".

Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>
Change-Id: I0b6e487f6b7fb3f56cf807a3c9b7bf910f51f38c