tfa-next: extend boot test coverage
Right now, we only check that RF-A boots up to the point where it prints
the "Page table activated." message on the UART. A few months back, this
used to be the last message printed by RF-A at boot time so this allowed
to check that RF-A successfully booted up all the way through. Since
then, development has continued and RF-A does a lot more stuff past that
point.

Update the expect scripts to wait for the last message printed by RF-A
today. This allows the CI to catch potential crashes happening between
the page table activation and the new stuff that is now executed after
that.

Both the files under expect/ and expect-lava/ directories need to be
updated. The latter is so that LAVA does not prematurely kill the FVP
before seeing the last message, which could truncate the output and
produce false negatives. The former is so that developers can reproduce
the same test setup on their local machine through the `run_local_ci.sh`
script.

Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>
Change-Id: I6e22742d67af1e5896f2d0dce64ea8a07fb6f41b