fix(ci): don't add a suffix to DOWNLOAD_SERVER_URL twice
The host_env will be sourced each time utils.sh is sourced. But utils.sh
is sourced in practically every script. When these scripts source each
other, we end up with multiple executions of host_env, which results in
the "tf-a" suffix being appended twice. Change this to always expand
from the source to avoid this.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I50d22cbc31b63e9f361a08ff7b21d2dc7c544f98