tfa-next: build with NEED_BL31=no
We are currently using tf-a's build system to build bl1, bl2, fip.
We are afterwards updating the fip with the Rust bl31 that we build
with rf-a's build system.

We should therefore avoid build C tf-a's bl31 as we don't need it.

With the current build command, we are currently building c tf-a's
bl31 when we build the fip.

 * Update TFA_FLAGS to avoid building c tf-a's bl31 when building
   the fip.

Change-Id: I44eb16db25eb1ca0f54a6712d31e61b6d16ca1d4
Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>