feat(status-bot): report failed subjobs
It's useful to know exactly which builds caused the daily to fail, not
just the group. Anyone looking at the report will dig in to find this
information if it's not manually inserted. So fetch all subjobs to find
every single failed job.

The spirit of the original script is kept, but the classes are fully
refactored since they didn't support true recursion. Processing is
streamlined a little by separating finding of the top level build from
the recursive fetching of build information and their children.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I0d3d845ff93366ce9bf3e1cad3a66f106551d60c