feat(ci-bot): run asynchronously and as a library
Fetching a daily with many failures takes a long time as each job has to
be fetched in sequence. Since we're dealing with a tree of jobs, there
is a lot of room for some of this to happen concurrently. Use asyncio
and aiohttp to achieve this.

Also allow the script to be run as a library and add commandline
arguments when running standalone.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I1bb9323de82675299181646881f4c05b074bc87d