feat(ci-bot): add a patch counter script
Add a simple script to count the number of patches in each project on
tforg from a given query. By default this will be the patches in review,
although this is configurable from an argument.

The implementation is very similar to the ci_status_bot so that they can
be used together from a script as well as standalone if necessary.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I982ea6894b89b72176c8473d52e46bb5d8a9f344