feat(ci-bot): add a script to send a daily CI report to Discord
The script uses the ci_status_bot and patch_count scripts to produce a
pair of messages that it sends to a provided Discord webhook. Since
Discord has a message limit of 2000 characters it truncates the result
if it would exceed it. The status is split into two messages so reduce
variability and to increase the number of failures that can be listed.
Finally, some error messages are printed to alert of anything going
wrong instead of failing silently.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: Ie2ff90f3fb5bd9e87dbfcb4673a8372fe3ef167e