feat(ci-bot): add config to run the discord notification bot every day
We want daily updates of the CI's and Gerrit's status. Add a systemd
service and timer to do this automatically. Wrap the discord script in
uv to reduce the required setup. Add some instructions on how to install
on your machine too, kept in the script's docstring to make it self
contained.

Signed-off-by: Boyan Karatotev <boyan.karatotev@arm.com>
Change-Id: I4b960576a78b84b909d167f9aa5975e44978ad71