tfa-next: introduce project config for copyright checks
When it comes to checking copyright and license headers, TF-A and RF-A
projects have different requirements. The script already handles
differences on the copyright format but there are more things to
consider, like the valid file extensions (C vs Rust) and the list of
directories and files to ignore.

Introduce a project configuration class to hold all these project
requirements. This is a more scalable and less intrusive approach than
relying on the '--rusted' argument everywhere in the script.

For now, in the intent to keep the patch diff simple to review, RF-A
project configuration only differs from TF-A's on the copyright format,
like it was the case before this patch. A subsequent patch will tune the
valid file extensions and the list of directories and files to ignore.

Signed-off-by: Sandrine Afsa <sandrine.afsa@arm.com>
Change-Id: I16d2930056e530b98932761827b75585a84b5001